third_party/blink/web_tests/http/tests/security/cross-frame-access-location-put-expected.txt

SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------

----- tests for putting window.history and its properties -----

PASS: window.location.assign should be 'function assign() { [native code] }' and is.
PASS: window.location.replace should be 'function replace() { [native code] }' and is.
PASS: window.location.reload should be 'function reload() { [native code] }' and is.
PASS: window.location.toString should be 'function toString() { [native code] }' and is.
PASS: window.location.customAttribute should be 'customAttribute' and is.
